+++ CMakeLists.txt
@@ -443,6 +443,7 @@ set(THIRDAI_SOURCES
     data/src/transformations/ner/NerTokenFromStringArray.cc
     data/src/transformations/ner/NerTokenizationUnigram.cc
     data/src/transformations/ner/NerDyadicDataProcessor.cc
+    data/src/transformations/ner/utils/TagTracker.cc
     data/src/transformations/ner/utils/TokenLabelCounter.cc
     data/src/transformations/ner/utils/utils.cc
     data/src/transformations/ner/rules/Rule.cc
