+++ data/tests/CMakeLists.txt
@@ -18,7 +18,9 @@ set(DATA_TESTS
     StateSerializationTest
     TextCompatTest
     MachMemorySerializationTest
-    NerRuleTests)
+    NerRuleTests
+    NerTagTests
+    NerTagTrackerTests)
 
 foreach(DATA_TEST IN LISTS DATA_TESTS)
   add_executable(${DATA_TEST} ${DATA_TEST}.cc)
